chore: further cleanup to facilitate testing
