+++ demos/leuvenshtein/src/algorithm/myers.rs
@@ -25,8 +25,13 @@
 
 use std::mem;
 
-use crate::data;
+use crate::data::*;
 use crate::enc_struct::EncStruct;
+use crate::util::{
+    self, apply_lookup_table_packed, unchecked_add_packed, unchecked_add_packed_assign,
+    unchecked_scalar_add_packed, unchecked_scalar_add_packed_assign, unchecked_scalar_mul_packed,
+    unchecked_scalar_mul_packed_assign, unchecked_sub_packed,
+};
 use std::collections::HashMap;
 use std::time::Instant;
 
@@ -139,6 +144,140 @@ pub fn print_matrix<T: std::fmt::Display>(matrix: &[Vec<T>], name: &str) {
     }
 }
 
+/// Plain Processing
+pub fn process_plain_query_enc_db(enc_struct: &mut EncStruct) {
+    // Get the min and max lenght of the db strings
+    let qlen = enc_struct.query.len() + 1;
+
+    // DB processing and encrypting
+    let mut db_len: HashMap<usize, usize> = HashMap::with_capacity(enc_struct.db_size);
+
+    for i in 0..enc_struct.db_size {
+        db_len.insert(i, NAME_LIST[i].len());
+    }
+
+    let db_max_size = db_len.values().max().unwrap();
+
+    // Max factor is defined as the size of the D matrix! (db_max_size + 1)
+    let mut max_factor = std::cmp::max(*db_max_size, qlen - 1);
+    max_factor += 1;
+
+    let th: usize = ((max_factor as f64) / 2.0).ceil() as usize;
+
+    let zero_enc = enc_struct.cks.encrypt(0u64);
+
+    // Build and fill all the h_matrices
+    let mut h_matrices: Vec<Vec<Vec<Ciphertext>>> = Vec::with_capacity(enc_struct.db_size);
+    let mut v_matrices: Vec<Vec<Vec<Ciphertext>>> = Vec::with_capacity(enc_struct.db_size);
+
+    for _ in 0..enc_struct.db_size {
+        let mut h_matrix: Vec<Vec<Ciphertext>> = Vec::with_capacity(max_factor);
+        let mut v_matrix: Vec<Vec<Ciphertext>> = Vec::with_capacity(max_factor);
+
+        for _ in 0..max_factor {
+            let mut vec: Vec<Ciphertext> = Vec::with_capacity(max_factor);
+            for _ in 0..max_factor {
+                vec.push(zero_enc.clone());
+            }
+            h_matrix.push(vec.clone());
+            v_matrix.push(vec);
+        }
+
+        for i in 0..max_factor {
+            v_matrix[i][0] = enc_struct.cks.encrypt(1u64);
+        }
+        for i in 0..max_factor {
+            h_matrix[0][i] = enc_struct.cks.encrypt(1u64);
+        }
+
+        h_matrices.push(h_matrix);
+        v_matrices.push(v_matrix);
+    }
+
+    let lut_min_vec_def = [0u64, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0].to_vec();
+    let lut_min = enc_struct
+        .sks
+        .generate_lookup_table_from_vector(&lut_min_vec_def);
+    let lut_min_vec = vec![lut_min; enc_struct.db_size];
+
+    let fpga_lut_single = LookupVector::new(lut_min_vec_def);
+    let lut_min_vec_fpga = vec![fpga_lut_single; enc_struct.db_size];
+
+    enc_struct.max_factor = max_factor;
+    enc_struct.th = th;
+    enc_struct.v_matrices = v_matrices;
+    enc_struct.h_matrices = h_matrices;
+    enc_struct.lut_min_vec_sw = lut_min_vec;
+    enc_struct.lut_min_vec_fpga = lut_min_vec_fpga;
+
+    enc_struct.time = Instant::now();
+}
+
+pub fn process_plain_part_i(index: usize, enc_struct: &mut EncStruct, fpga_enable: bool) {
+    let query_padded = enc_struct.query.pad_to_width(enc_struct.max_factor - 1);
+
+    for j in 1..enc_struct.max_factor {
+        if usize::abs_diff(index, j) <= enc_struct.th {
+            let eq: Vec<Ciphertext> = util::get_db_enc_vec(
+                query_padded.chars().nth(index - 1).unwrap(),
+                j - 1,
+                &enc_struct.db_enc_map,
+            );
+
+            let vin = util::extract_number_elements(&enc_struct.v_matrices, index, j - 1);
+            let hin = util::extract_number_elements(&enc_struct.h_matrices, index - 1, j);
+
+            let v1 = unchecked_scalar_add_packed(&enc_struct.sks, vin.iter().collect(), 1);
+            let h1 = unchecked_scalar_add_packed(&enc_struct.sks, hin.iter().collect(), 1);
+
+            let key1 = unchecked_scalar_mul_packed(&enc_struct.sks, h1.iter().collect(), 3);
+
+            let key12 =
+                unchecked_add_packed(&enc_struct.sks, key1.iter().collect(), eq.iter().collect());
+
+            let key =
+                unchecked_add_packed(&enc_struct.sks, key12.iter().collect(), v1.iter().collect());
+
+            let mut ct_res: Vec<Ciphertext>;
+            if fpga_enable {
+                ct_res = key.clone();
+
+                #[cfg(feature = "fpga")]
+                {
+                    let ref_lookupvector: [&LookupVector; LUT_SIZE] =
+                        std::array::from_fn(|i| &enc_struct.lut_min_vec_fpga[i]);
+                    enc_struct
+                        .fpga_key
+                        .apply_lookup_vector_packed_assign(&mut ct_res, &ref_lookupvector);
+                }
+            } else {
+                ct_res = apply_lookup_table_packed(
+                    &enc_struct.sks,
+                    key.iter().collect(),
+                    &enc_struct.lut_min_vec_sw,
+                );
+            }
+
+            unchecked_scalar_add_packed_assign(&enc_struct.sks, &mut ct_res, 16);
+
+            let v_res = unchecked_sub_packed(
+                &enc_struct.sks,
+                ct_res.iter().collect(),
+                hin.iter().collect(),
+            );
+            let h_res = unchecked_sub_packed(
+                &enc_struct.sks,
+                ct_res.iter().collect(),
+                vin.iter().collect(),
+            );
+
+            util::write_number_elements(&mut enc_struct.v_matrices, &v_res, index, j);
+            util::write_number_elements(&mut enc_struct.h_matrices, &h_res, index, j);
+        }
+    }
+}
+
+/// Encrypted Processing
 pub fn process_enc_query_enc_db(enc_struct: &mut EncStruct) {
     // Get the min and max lenght of the db strings
     let qlen = enc_struct.query.len() + 1;
@@ -148,7 +287,7 @@ pub fn process_enc_query_enc_db(enc_struct: &mut EncStruct) {
     let mut db_len: HashMap<usize, usize> = HashMap::with_capacity(enc_struct.db_size);
 
     for i in 0..enc_struct.db_size {
-        db_len.insert(i, data::NAME_LIST[i].len());
+        db_len.insert(i, NAME_LIST[i].len());
     }
 
     let db_max_size = *db_len.values().max().unwrap();
@@ -166,12 +305,12 @@ pub fn process_enc_query_enc_db(enc_struct: &mut EncStruct) {
     let q_enc = query_padded
         .bytes() // convert char to int
         .map(|c| enc_struct.cks.encrypt((c - scale_factor) as u64))
-        .collect::<Vec<tfhe::shortint::Ciphertext>>();
+        .collect::<Vec<Ciphertext>>();
 
     let q2_enc = query_padded
         .bytes() // convert char to int
         .map(|c| enc_struct.cks.encrypt(((c - scale_factor) >> 4) as u64))
-        .collect::<Vec<tfhe::shortint::Ciphertext>>();
+        .collect::<Vec<Ciphertext>>();
 
     let zero_enc = enc_struct.cks.encrypt(0u64);
     let one_enc = enc_struct.cks.encrypt(1u64);
@@ -180,17 +319,17 @@ pub fn process_enc_query_enc_db(enc_struct: &mut EncStruct) {
     let mut db1_enc_matrix: Vec<Vec<Ciphertext>> = Vec::with_capacity(enc_struct.db_size);
 
     for i in 0..enc_struct.db_size {
-        let padded_name = data::NAME_LIST[i].pad_to_width(max_factor - 1);
+        let padded_name = NAME_LIST[i].pad_to_width(max_factor - 1);
 
         let name_enc = padded_name
             .bytes() // convert char to int
             .map(|c| enc_struct.cks.encrypt((c - scale_factor) as u64)) // Encrypts
-            .collect::<Vec<tfhe::shortint::Ciphertext>>();
+            .collect::<Vec<Ciphertext>>();
 
         let name1_enc = padded_name
             .bytes() // convert char to int
             .map(|c| enc_struct.cks.encrypt(((c - scale_factor) >> 4) as u64))
-            .collect::<Vec<tfhe::shortint::Ciphertext>>();
+            .collect::<Vec<Ciphertext>>();
 
         db_enc_matrix.push(name_enc);
         db1_enc_matrix.push(name1_enc);
@@ -223,17 +362,15 @@ pub fn process_enc_query_enc_db(enc_struct: &mut EncStruct) {
     let lut_min_vec_fpga = vec![lut_min_fpga; enc_struct.db_size];
 
     // Build and fill all the h_matrices
-    let mut h_matrices: Vec<Vec<Vec<tfhe::shortint::Ciphertext>>> =
-        Vec::with_capacity(enc_struct.db_size);
-    let mut v_matrices: Vec<Vec<Vec<tfhe::shortint::Ciphertext>>> =
-        Vec::with_capacity(enc_struct.db_size);
+    let mut h_matrices: Vec<Vec<Vec<Ciphertext>>> = Vec::with_capacity(enc_struct.db_size);
+    let mut v_matrices: Vec<Vec<Vec<Ciphertext>>> = Vec::with_capacity(enc_struct.db_size);
 
     for _ in 0..enc_struct.db_size {
-        let mut h_matrix: Vec<Vec<tfhe::shortint::Ciphertext>> = Vec::with_capacity(max_factor);
-        let mut v_matrix: Vec<Vec<tfhe::shortint::Ciphertext>> = Vec::with_capacity(max_factor);
+        let mut h_matrix: Vec<Vec<Ciphertext>> = Vec::with_capacity(max_factor);
+        let mut v_matrix: Vec<Vec<Ciphertext>> = Vec::with_capacity(max_factor);
 
         for _ in 0..max_factor {
-            let mut vec: Vec<tfhe::shortint::Ciphertext> = Vec::with_capacity(max_factor);
+            let mut vec: Vec<Ciphertext> = Vec::with_capacity(max_factor);
             for _ in 0..max_factor {
                 vec.push(zero_enc.clone());
             }
@@ -271,6 +408,145 @@ pub fn process_enc_query_enc_db(enc_struct: &mut EncStruct) {
     enc_struct.lut_min_vec_fpga = lut_min_vec_fpga;
 
     enc_struct.time = Instant::now();
+}
+
+pub fn process_part_i(index: usize, enc_struct: &mut EncStruct, fpga_enable: bool) {
+    let q1_vec: Vec<tfhe::shortint::prelude::Ciphertext> =
+        vec![enc_struct.q_enc[index - 1].clone(); enc_struct.db_size];
+    let q2_vec: Vec<tfhe::shortint::prelude::Ciphertext> =
+        vec![enc_struct.q2_enc[index - 1].clone(); enc_struct.db_size];
+
+    let one_enc_vec_ref: Vec<&Ciphertext> = enc_struct.one_enc_vec.iter().collect();
+
+    for j in 1..enc_struct.max_factor {
+        if usize::abs_diff(index, j) <= enc_struct.th {
+            // Check the first part of the character
+            let mut eq1 = unchecked_sub_packed(
+                &enc_struct.sks,
+                q1_vec.iter().collect(),
+                util::get_column(&enc_struct.db_enc_matrix, j - 1)
+                    .iter()
+                    .collect(),
+            );
+
+            unchecked_scalar_add_packed_assign(&enc_struct.sks, &mut eq1, 16);
+
+            let mut eq1_lut = Vec::new();
+
+            if fpga_enable {
+                eq1_lut = eq1.clone();
+                #[cfg(feature = "fpga")]
+                {
+                    let ref_lookupvector: [&LookupVector; LUT_SIZE] =
+                        std::array::from_fn(|i| &enc_struct.lut_1eq_vec_fpga[i]);
+                    enc_struct
+                        .fpga_key
+                        .apply_lookup_vector_packed_assign(&mut eq1_lut, &ref_lookupvector);
+                }
+            } else {
+                let ct = apply_lookup_table_packed(
+                    &enc_struct.sks,
+                    eq1.iter().collect(),
+                    &enc_struct.lut_1eq_vec_sw,
+                );
+                eq1_lut.extend(ct);
+            }
+
+            let eq1_ref: Vec<&Ciphertext> = eq1_lut.iter().collect();
+
+            eq1 = unchecked_sub_packed(&enc_struct.sks, one_enc_vec_ref.clone(), eq1_ref);
+
+            unchecked_scalar_add_packed_assign(&enc_struct.sks, &mut eq1, 16);
+
+            let mut eq2 = unchecked_sub_packed(
+                &enc_struct.sks,
+                q2_vec.iter().collect(),
+                util::get_column(&enc_struct.db1_enc_matrix, j - 1)
+                    .iter()
+                    .collect(),
+            );
+
+            unchecked_scalar_mul_packed_assign(&enc_struct.sks, &mut eq2, 2);
+
+            unchecked_add_packed_assign(&enc_struct.sks, &mut eq2, eq1.iter().collect());
+
+            let mut eq2_lut = Vec::new();
+
+            if fpga_enable {
+                eq2_lut = eq2.clone();
+                #[cfg(feature = "fpga")]
+                {
+                    let ref_lookupvector: [&LookupVector; LUT_SIZE] =
+                        std::array::from_fn(|i| &enc_struct.lut_eq_vec_fpga[i]);
+                    enc_struct
+                        .fpga_key
+                        .apply_lookup_vector_packed_assign(&mut eq2_lut, &ref_lookupvector);
+                }
+            } else {
+                let ct = apply_lookup_table_packed(
+                    &enc_struct.sks,
+                    eq2.iter().collect(),
+                    &enc_struct.lut_eq_vec_sw,
+                );
+                eq2_lut.extend(ct);
+            }
+
+            let vin = util::extract_number_elements(&enc_struct.v_matrices, index, j - 1);
+            let hin = util::extract_number_elements(&enc_struct.h_matrices, index - 1, j);
+
+            let v1 = unchecked_scalar_add_packed(&enc_struct.sks, vin.iter().collect(), 1);
+            let h1 = unchecked_scalar_add_packed(&enc_struct.sks, hin.iter().collect(), 1);
+
+            let key1 = unchecked_scalar_mul_packed(&enc_struct.sks, h1.iter().collect(), 3);
+
+            let key12 = unchecked_add_packed(
+                &enc_struct.sks,
+                key1.iter().collect(),
+                eq2_lut.iter().collect(),
+            );
+
+            let key =
+                unchecked_add_packed(&enc_struct.sks, key12.iter().collect(), v1.iter().collect());
+
+            let mut ct_res = Vec::new();
+
+            if fpga_enable {
+                ct_res = key.clone();
+
+                #[cfg(feature = "fpga")]
+                {
+                    let ref_lookupvector: [&LookupVector; LUT_SIZE] =
+                        std::array::from_fn(|i| &enc_struct.lut_min_vec_fpga[i]);
+                    enc_struct
+                        .fpga_key
+                        .apply_lookup_vector_packed_assign(&mut ct_res, &ref_lookupvector);
+                }
+            } else {
+                let ct = apply_lookup_table_packed(
+                    &enc_struct.sks,
+                    key.iter().collect(),
+                    &enc_struct.lut_min_vec_sw,
+                );
+                ct_res.extend(ct);
+            }
+
+            unchecked_scalar_add_packed_assign(&enc_struct.sks, &mut ct_res, 16);
+
+            let v_res = unchecked_sub_packed(
+                &enc_struct.sks,
+                ct_res.iter().collect(),
+                hin.iter().collect(),
+            );
+            let h_res = unchecked_sub_packed(
+                &enc_struct.sks,
+                ct_res.iter().collect(),
+                vin.iter().collect(),
+            );
+
+            util::write_number_elements(&mut enc_struct.v_matrices, &v_res, index, j);
+            util::write_number_elements(&mut enc_struct.h_matrices, &h_res, index, j);
+        }
+    }
 }
 
 #[cfg(test)]
